refactor: systematically remove as any usage from test files
- Extracted JWT validation logic from TokenManager to separate jwt.ts utility
- Created comprehensive JWT utility tests in src/utils/__tests__/jwt.test.ts
- Updated TokenManager to use the new JWT utility function
- Added resetForTesting() method to TokenManager for proper test isolation
- Removed all JWT validation tests from token-manager.test.ts to avoid duplication
- Fixed token manager test setup to prevent rate limiting interference
- Verified all 1123 tests pass after changes

This eliminates the last remaining as any usage in the codebase while maintaining
comprehensive test coverage for JWT token validation functionality.
